Align backend package with gxj namespace and refresh import summary
