+++ demo/pom.xml
@@ -8,11 +8,11 @@
         <version>3.5.6</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>com.example</groupId>
-    <artifactId>demo</artifactId>
+    <groupId>com.gxj</groupId>
+    <artifactId>crop-yield</artifactId>
     <version>0.0.1-SNAPSHOT</version>
-    <name>demo</name>
-    <description>demo</description>
+    <name>crop-yield</name>
+    <description>Yunnan crop yield analytics backend</description>
     <url/>
     <licenses>
         <license/>
