--- demo/src/main/java/com/example/demo/crop/CropRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.crop;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CropRepository extends JpaRepository<Crop, Long> {
-
-    Optional<Crop> findByNameIgnoreCase(String name);
-}
