--- demo/src/main/java/com/example/demo/crop/CropResponse.java
@@ -1,9 +1,0 @@
-package com.example.demo.crop;
-
-public record CropResponse(
-        Long id,
-        String name,
-        String category,
-        String description
-) {
-}
