--- demo/src/main/java/com/example/demo/dashboard/DashboardController.java
@@ -1,22 +1,0 @@
-package com.example.demo.dashboard;
-
-import com.example.demo.dashboard.dto.DashboardSummaryResponse;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api/dashboard")
-public class DashboardController {
-
-    private final DashboardService dashboardService;
-
-    public DashboardController(DashboardService dashboardService) {
-        this.dashboardService = dashboardService;
-    }
-
-    @GetMapping("/summary")
-    public DashboardSummaryResponse getSummary() {
-        return dashboardService.getSummary();
-    }
-}
