--- demo/src/main/java/com/example/demo/dashboard/dto/CropShare.java
@@ -1,9 +1,0 @@
-package com.example.demo.dashboard.dto;
-
-public record CropShare(
-        String cropName,
-        double production,
-        double sownArea,
-        double share
-) {
-}
