--- demo/src/main/java/com/example/demo/dashboard/dto/DashboardSummaryResponse.java
@@ -1,16 +1,0 @@
-package com.example.demo.dashboard.dto;
-
-import java.util.List;
-
-public record DashboardSummaryResponse(
-        double totalProduction,
-        double totalSownArea,
-        double averageYield,
-        long recordCount,
-        List<TrendPoint> productionTrend,
-        List<CropShare> cropStructure,
-        List<RegionProductionSummary> regionComparisons,
-        List<RecentYieldRecord> recentRecords,
-        List<ForecastPoint> forecastOutlook
-) {
-}
