--- demo/src/main/java/com/example/demo/dashboard/dto/ForecastPoint.java
@@ -1,10 +1,0 @@
-package com.example.demo.dashboard.dto;
-
-public record ForecastPoint(
-        String label,
-        double value,
-        double lowerBound,
-        double upperBound,
-        String model
-) {
-}
