--- demo/src/main/java/com/example/demo/dashboard/dto/RegionProductionSummary.java
@@ -1,10 +1,0 @@
-package com.example.demo.dashboard.dto;
-
-public record RegionProductionSummary(
-        String regionName,
-        String level,
-        double production,
-        double sownArea,
-        double yieldPerHectare
-) {
-}
