--- demo/src/main/java/com/example/demo/dashboard/dto/TrendPoint.java
@@ -1,7 +1,0 @@
-package com.example.demo.dashboard.dto;
-
-public record TrendPoint(
-        String label,
-        double value
-) {
-}
