--- demo/src/main/java/com/example/demo/datamanagement/DataImportController.java
@@ -1,24 +1,0 @@
-package com.example.demo.datamanagement;
-
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
-
-@RestController
-@RequestMapping("/api/data-import")
-public class DataImportController {
-
-    private final DataImportService dataImportService;
-
-    public DataImportController(DataImportService dataImportService) {
-        this.dataImportService = dataImportService;
-    }
-
-    @PostMapping(value = "/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public DataImportResponse upload(@RequestPart("file") MultipartFile file) {
-        return dataImportService.importData(file);
-    }
-}
