--- demo/src/main/java/com/example/demo/datamanagement/DataImportResponse.java
@@ -1,15 +1,0 @@
-package com.example.demo.datamanagement;
-
-import com.example.demo.yielddata.YieldRecordResponse;
-
-import java.util.List;
-
-public record DataImportResponse(
-        int totalRows,
-        int insertedRows,
-        int updatedRows,
-        int skippedRows,
-        List<String> warnings,
-        List<YieldRecordResponse> preview
-) {
-}
