--- demo/src/main/java/com/example/demo/region/RegionController.java
@@ -1,31 +1,0 @@
-package com.example.demo.region;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/regions")
-public class RegionController {
-
-    private final RegionRepository regionRepository;
-
-    public RegionController(RegionRepository regionRepository) {
-        this.regionRepository = regionRepository;
-    }
-
-    @GetMapping
-    public List<RegionResponse> listAll() {
-        return regionRepository.findAll().stream()
-                .map(region -> new RegionResponse(
-                        region.getId(),
-                        region.getName(),
-                        region.getLevel(),
-                        region.getParentName(),
-                        region.getDescription()
-                ))
-                .toList();
-    }
-}
