--- demo/src/main/java/com/example/demo/region/RegionRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.region;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RegionRepository extends JpaRepository<Region, Long> {
-
-    Optional<Region> findByNameIgnoreCase(String name);
-}
