--- demo/src/main/java/com/example/demo/region/RegionResponse.java
@@ -1,10 +1,0 @@
-package com.example.demo.region;
-
-public record RegionResponse(
-        Long id,
-        String name,
-        String level,
-        String parentName,
-        String description
-) {
-}
