+++ demo/src/main/java/com/gxj/cropyield/CropYieldApplication.java
@@ -1,0 +1,13 @@
+package com.gxj.cropyield;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CropYieldApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CropYieldApplication.class, args);
+    }
+
+}
