+++ demo/src/main/java/com/gxj/cropyield/crop/Crop.java
@@ -1,0 +1,51 @@
+package com.gxj.cropyield.crop;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "crops")
+public class Crop {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String name;
+
+    @Column(nullable = false)
+    private String category;
+
+    @Column(length = 1000)
+    private String description;
+
+    protected Crop() {
+    }
+
+    public Crop(String name, String category, String description) {
+        this.name = name;
+        this.category = category;
+        this.description = description;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getCategory() {
+        return category;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+}
