+++ demo/src/main/java/com/gxj/cropyield/crop/CropController.java
@@ -1,0 +1,30 @@
+package com.gxj.cropyield.crop;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/crops")
+public class CropController {
+
+    private final CropRepository cropRepository;
+
+    public CropController(CropRepository cropRepository) {
+        this.cropRepository = cropRepository;
+    }
+
+    @GetMapping
+    public List<CropResponse> listAll() {
+        return cropRepository.findAll().stream()
+                .map(crop -> new CropResponse(
+                        crop.getId(),
+                        crop.getName(),
+                        crop.getCategory(),
+                        crop.getDescription()
+                ))
+                .toList();
+    }
+}
