+++ demo/src/main/java/com/gxj/cropyield/crop/CropRepository.java
@@ -1,0 +1,10 @@
+package com.gxj.cropyield.crop;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface CropRepository extends JpaRepository<Crop, Long> {
+
+    Optional<Crop> findByNameIgnoreCase(String name);
+}
