+++ demo/src/main/java/com/gxj/cropyield/crop/CropResponse.java
@@ -1,0 +1,9 @@
+package com.gxj.cropyield.crop;
+
+public record CropResponse(
+        Long id,
+        String name,
+        String category,
+        String description
+) {
+}
