+++ demo/src/main/java/com/gxj/cropyield/dashboard/DashboardController.java
@@ -1,0 +1,22 @@
+package com.gxj.cropyield.dashboard;
+
+import com.gxj.cropyield.dashboard.dto.DashboardSummaryResponse;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/dashboard")
+public class DashboardController {
+
+    private final DashboardService dashboardService;
+
+    public DashboardController(DashboardService dashboardService) {
+        this.dashboardService = dashboardService;
+    }
+
+    @GetMapping("/summary")
+    public DashboardSummaryResponse getSummary() {
+        return dashboardService.getSummary();
+    }
+}
