+++ demo/src/main/java/com/gxj/cropyield/dashboard/dto/CropShare.java
@@ -1,0 +1,9 @@
+package com.gxj.cropyield.dashboard.dto;
+
+public record CropShare(
+        String cropName,
+        double production,
+        double sownArea,
+        double share
+) {
+}
