+++ demo/src/main/java/com/gxj/cropyield/dashboard/dto/ForecastPoint.java
@@ -1,0 +1,10 @@
+package com.gxj.cropyield.dashboard.dto;
+
+public record ForecastPoint(
+        String label,
+        double value,
+        double lowerBound,
+        double upperBound,
+        String model
+) {
+}
