+++ demo/src/main/java/com/gxj/cropyield/dashboard/dto/RecentYieldRecord.java
@@ -1,0 +1,16 @@
+package com.gxj.cropyield.dashboard.dto;
+
+import java.time.LocalDate;
+
+public record RecentYieldRecord(
+        String cropName,
+        String regionName,
+        int year,
+        double sownArea,
+        double production,
+        double yieldPerHectare,
+        Double averagePrice,
+        double estimatedRevenue,
+        LocalDate collectedAt
+) {
+}
