+++ demo/src/main/java/com/gxj/cropyield/dashboard/dto/RegionProductionSummary.java
@@ -1,0 +1,10 @@
+package com.gxj.cropyield.dashboard.dto;
+
+public record RegionProductionSummary(
+        String regionName,
+        String level,
+        double production,
+        double sownArea,
+        double yieldPerHectare
+) {
+}
