+++ demo/src/main/java/com/gxj/cropyield/dashboard/dto/TrendPoint.java
@@ -1,0 +1,7 @@
+package com.gxj.cropyield.dashboard.dto;
+
+public record TrendPoint(
+        String label,
+        double value
+) {
+}
