+++ demo/src/main/java/com/gxj/cropyield/datamanagement/DataImportResponse.java
@@ -1,0 +1,15 @@
+package com.gxj.cropyield.datamanagement;
+
+import com.gxj.cropyield.yielddata.YieldRecordResponse;
+
+import java.util.List;
+
+public record DataImportResponse(
+        int totalRows,
+        int insertedRows,
+        int updatedRows,
+        int skippedRows,
+        List<String> warnings,
+        List<YieldRecordResponse> preview
+) {
+}
