+++ demo/src/main/java/com/gxj/cropyield/region/RegionRepository.java
@@ -1,0 +1,10 @@
+package com.gxj.cropyield.region;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RegionRepository extends JpaRepository<Region, Long> {
+
+    Optional<Region> findByNameIgnoreCase(String name);
+}
