+++ demo/src/main/java/com/gxj/cropyield/region/RegionResponse.java
@@ -1,0 +1,10 @@
+package com.gxj.cropyield.region;
+
+public record RegionResponse(
+        Long id,
+        String name,
+        String level,
+        String parentName,
+        String description
+) {
+}
