+++ demo/src/test/java/com/gxj/cropyield/CropYieldApplicationTests.java
@@ -1,0 +1,12 @@
+package com.gxj.cropyield;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CropYieldApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+}
